+++ social-media-react/src/components/authentication/RegistrationForm.jsx
