+++ social-media-react/src/components/authentication/__tests__/RegistrationForm.test.js
