+++ social-media-react/src/pages/Registration.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import RegistrationForm from "../components/forms/RegistrationForm";
+import RegistrationForm from "../components/authentication/RegistrationForm";
 
 function Registration() {
   return (
